@@ -1,32 +1,53 @@
 {
   "name": "whatwg-url",
-  "version": "5.0.0",
+  "version": "14.2.0",
   "description": "An implementation of the WHATWG URL Standard's URL API and parsing machinery",
-  "main": "lib/public-api.js",
+  "main": "index.js",
   "files": [
-    "lib/"
+    "index.js",
+    "webidl2js-wrapper.js",
+    "lib/*.js"
   ],
   "author": "Sebastian Mayr <github@smayr.name>",
   "license": "MIT",
   "repository": "jsdom/whatwg-url",
   "dependencies": {
-    "tr46": "~0.0.3",
-    "webidl-conversions": "^3.0.0"
+    "tr46": "^5.1.0",
+    "webidl-conversions": "^7.0.0"
   },
   "devDependencies": {
-    "eslint": "^2.6.0",
-    "istanbul": "~0.4.3",
-    "mocha": "^2.2.4",
-    "recast": "~0.10.29",
-    "request": "^2.55.0",
-    "webidl2js": "^3.0.2"
+    "@domenic/eslint-config": "^4.0.1",
+    "benchmark": "^2.1.4",
+    "c8": "^10.1.3",
+    "esbuild": "^0.25.1",
+    "eslint": "^9.22.0",
+    "globals": "^16.0.0",
+    "webidl2js": "^18.0.0"
+  },
+  "engines": {
+    "node": ">=18"
   },
   "scripts": {
-    "build": "node scripts/transform.js && node scripts/convert-idl.js",
-    "coverage": "istanbul cover node_modules/mocha/bin/_mocha",
-    "lint": "eslint .",
-    "prepublish": "npm run build",
-    "pretest": "node scripts/get-latest-platform-tests.js && npm run build",
-    "test": "mocha"
+    "coverage": "c8 node --test --experimental-test-coverage test/*.js",
+    "lint": "eslint",
+    "prepare": "node scripts/transform.js",
+    "pretest": "node scripts/get-latest-platform-tests.js && node scripts/transform.js",
+    "build-live-viewer": "esbuild --bundle --format=esm --sourcemap --outfile=live-viewer/whatwg-url.mjs index.js",
+    "test": "node --test test/*.js",
+    "bench": "node scripts/benchmark.js"
+  },
+  "c8": {
+    "reporter": [
+      "text",
+      "html"
+    ],
+    "exclude": [
+      "lib/Function.js",
+      "lib/URL.js",
+      "lib/URLSearchParams.js",
+      "lib/utils.js",
+      "scripts/",
+      "test/"
+    ]
   }
 }
